Improve application performance and developer experience with several updates
Installs nanoid, adds '@assets' path alias to tsconfig.json, adds database indexes to projects, resources, and timesheet_entries tables, and enforces integer type for completionPercentage in shared/routes.ts.

+++ shared/routes.ts
@@ -133,7 +133,7 @@ export const api = {
       method: 'PUT' as const,
       path: '/api/projects/:id',
       input: insertProjectSchema.partial().extend({
-        completionPercentage: z.number().optional(),
+        completionPercentage: z.number().int().optional(),
         health: z.string().optional(),
         startDate: z.union([z.string(), z.date(), z.null()]).optional(),
         endDate: z.union([z.string(), z.date(), z.null()]).optional(),
